[SERVER] log tweaks
